More accurate description for "Checking error codes"
Unfortunately the mechanism used to ensure it has no impact on build times
makes it *look* like it has a huge impact. Hopefully this change clarifies
what is going on. The issue is that ninja prints each task when it starts and
when it ends, but it keeps displaying the last-finished task until the next
task either starts or finishes. The error code checker is an (order-only)
dependency of linking so that it can happen in parallel with compiles, but
will still run every time any file changes. It will finish quickly, but the
linking can't start until all required compilations finish.